refactor: Update indentation and code style in .csharpierrc and .editorconfig
This commit updates the indentation and code style settings in the .csharpierrc and .editorconfig files. In the .csharpierrc file, the "printWidth" setting is changed to 100, "useTabs" is set to false, "tabWidth" is set to 2, and "endOfLine" is set to "auto". In the .editorconfig file, the "indent_size" setting is changed to 2. These changes improve the consistency and readability of the codebase.